requirements.txt
mysql-connector-python==8.0.32
python-dotenv==1.0.0
configparser==5.3.0
cryptography==41.0.1